moving branch up -- am going to try merge the unrelated pieces in batches, and then will move branch up again.

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/dht/db/AltLocFinder.java
@@ -295,7 +295,7 @@ public class AltLocFinder {
             }
             
             Set<? extends IpPort> proxies = pushProxies.getPushProxies();
-            int features = pushProxies.getFeatures();
+            byte features = pushProxies.getFeatures();
             int fwtVersion = pushProxies.getFwtVersion();
             IpPort ipp = new IpPortImpl(addr, pushProxies.getPort());
